fix(build_transformer): npm error when building without `package.json` (#3)
* Fix defi sdk example build

Support projects that require npm (komodo-wallet) and newer projects that don't (no package.json present)

* Overwrite existing files when extracting zips

* Add code validation and unit test workflows for PRs

* Revert "Add code validation and unit test workflows for PRs"

This reverts commit 526e17d5cc98a0280f40f339db0c710acf7be1a2.

+++ packages/komodo_wallet_build_transformer/lib/src/steps/fetch_defi_api_build_step.dart
@@ -320,6 +320,13 @@ class FetchDefiApiStep extends BuildStep {
   }
 
   Future<void> _updateWebPackages() async {
+    // First check for a `package.json` file in the root of the project
+    final packageJsonFile = File(path.join(artifactOutputPath, 'package.json'));
+    if (!packageJsonFile.existsSync()) {
+      _log.info('No package.json file found in $artifactOutputPath');
+      return;
+    }
+
     _log
       ..info('Updating Web platform...')
       ..fine('Running npm install in $artifactOutputPath');
@@ -478,9 +485,11 @@ class FetchDefiApiStep extends BuildStep {
     try {
       // Determine the platform to use the appropriate extraction command
       if (Platform.isMacOS || Platform.isLinux) {
-        // For macOS and Linux, use the `unzip` command
-        final result =
-            await Process.run('unzip', [zipFilePath, '-d', destinationFolder]);
+        // For macOS and Linux, use the `unzip` command with overwrite option
+        final result = await Process.run(
+          'unzip',
+          ['-o', zipFilePath, '-d', destinationFolder],
+        );
         if (result.exitCode != 0) {
           throw Exception('Error extracting zip file: ${result.stderr}');
         }
